Update MergeArray.cs
Just a little overhead with the usage of the `Queue<T>` object, but it was simpler than working directly with the input arrays. Should be 
a O(4n+4m) operation, where `n` is the number of items in `array1` and `m` is the number of items in `array2`. This is, by no means, the most performant solution.

Some individual performance notes:
* Instantiating the `Queue` with a collection is O(n) where n is the number of items in the input collection.
* Instantiating the `Queue` with a capacity is O(n) where n is the capacity.
* Calling `Queue.Count` is O(1)
* Calling `Queue.Enqueue` in which the current Count is less than than the capacity (because we instantiated it with a capacity) is O(1). Otherwise it's O(n) because it has to resize.
* Calling `Queue.Dequeue` is O(1)
* Calling `Queue.ToArray` is O(n) where n is the current Queue's Count